new: books, categories, login title (query)

+++ app/Http/Livewire/Pages/CategoriesComponent.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\Pages;
 
+use App\Models\Books;
 use App\Models\Categories;
 use Livewire\Component;
 
@@ -10,9 +11,14 @@ class CategoriesComponent extends Component
     public function destroy($categoriesId)
     {
         $findCategories = Categories::find($categoriesId);
-        $findCategories->delete();
-        session()->flash('message', 'Category ' . $findCategories->name . ' successfully deleted!');
-        return redirect()->back();
+        $checkBooks = Books::where("categories_id", $findCategories->id)->first();
+        if ($checkBooks) {
+            session()->flash('warning', 'Have a relation books');
+        } else {
+            $findCategories->delete();
+            session()->flash('message', 'Category ' . $findCategories->name . ' successfully deleted!');
+            return redirect()->back();
+        }
     }
     public function render()
     {
